+++ 5-period/java-bridge-main/src/main/java/bridge/controller/BridgeController.java
@@ -1,14 +1,14 @@
 package bridge.controller;
 
-import bridge.Bridge;
-import bridge.BridgeGame;
-import bridge.BridgeMaker;
-import bridge.Direction;
-import bridge.RestartCommand;
+import bridge.domain.bridge.Bridge;
+import bridge.domain.bridge.BridgeGame;
+import bridge.domain.bridge.BridgeMaker;
+import bridge.domain.Direction;
+import bridge.domain.RestartCommand;
 import bridge.dto.MoveResultDto;
 import bridge.dto.TotalResultDto;
 import bridge.exception.ExceptionHandler;
-import bridge.generator.BridgeRandomNumberGenerator;
+import bridge.domain.generator.BridgeRandomNumberGenerator;
 import bridge.view.InputView;
 import bridge.view.OutputView;
 import java.util.ArrayList;
