+++ 5-period/java-bridge-main/src/main/java/bridge/domain/Direction.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.domain;
 
 import bridge.exception.CustomIllegalArgumentException;
 import bridge.exception.ErrorMessage;
