+++ 5-period/java-bridge-main/src/main/java/bridge/domain/bridge/BridgeGame.java
@@ -1,4 +1,6 @@
-package bridge;
+package bridge.domain.bridge;
+
+import bridge.domain.Direction;
 
 /**
  * 다리 건너기 게임을 관리하는 클래스
