+++ 5-period/java-bridge-main/src/main/java/bridge/dto/MoveResultDto.java
@@ -1,6 +1,6 @@
 package bridge.dto;
 
-import bridge.Direction;
+import bridge.domain.Direction;
 
 public record MoveResultDto(boolean isRightMove, char direction) {
     public static MoveResultDto of(boolean isRightMove, Direction direction) {
