+++ 5-period/java-bridge-main/src/test/java/bridge/ApplicationTest.java
@@ -5,7 +5,8 @@ import static camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.util.Lists.newArrayList;
 
-import bridge.generator.BridgeNumberGenerator;
+import bridge.domain.bridge.BridgeMaker;
+import bridge.domain.generator.BridgeNumberGenerator;
 import camp.nextstep.edu.missionutils.test.NsTest;
 import java.util.List;
 import org.junit.jupiter.api.Test;
